+++ src/azure-cli-core/HISTORY.rst
@@ -7,6 +7,8 @@ Release History
 ++++++
 * Added limited support for positional arguments.
 * Fix issue where `--query` could not be used with `--ids`. [#5591](https://github.com/Azure/azure-cli/issues/5591)
+* Improves piping scenarios from commands when using `--ids`. Supports `-o tsv` with a query specified or `-o json`
+  without specifying a query.
 
 2.0.31
 ++++++